refactor(e2e): derive CLI artifact consumers
