+++ tools/e2e/prepare-e2e-workflow-boundary.mts
@@ -8,7 +8,7 @@ import { fileURLToPath } from "node:url";
 import { isDeepStrictEqual } from "node:util";
 import YAML from "yaml";
 import { SHARED_E2E_JOB_ID } from "./credential-free-tests.mts";
-import { E2E_ACTION_PROVENANCE, E2E_JOB_CONTRACT } from "./workflow-contract-registry.mts";
+import { E2E_ACTION_PROVENANCE, E2E_JOB_POLICY } from "./workflow-boundary-policy.mts";
 
 const REPO_ROOT = join(dirname(fileURLToPath(import.meta.url)), "..", "..");
 const DEFAULT_ACTION_PATH = join(REPO_ROOT, ".github", "actions", "prepare-e2e", "action.yaml");
@@ -19,13 +19,13 @@ export const PREPARE_E2E_ACTION = PREPARE_E2E_ACTION_PROVENANCE.reference;
 export const PREPARE_E2E_STEP = "Prepare E2E workspace";
 
 const CHECKOUT_LOCAL_PREPARE_E2E_ACTION = "./.github/actions/prepare-e2e";
-export const CLI_ARTIFACT_PRODUCER_JOB = E2E_JOB_CONTRACT.cliArtifactProducer;
+export const CLI_ARTIFACT_PRODUCER_JOB = E2E_JOB_POLICY.cliArtifactProducer;
 const PREINSTALLED_E2E_JOBS = new Set(["staging-brev-launchable"]);
 const RETIRED_SELECTOR_COMPATIBILITY_JOB = "retired-selector-compatibility";
 
-export const PREPARE_E2E_NO_BUILD_JOBS = new Set<string>(E2E_JOB_CONTRACT.prepareNoBuild);
+export const PREPARE_E2E_NO_BUILD_JOBS = new Set<string>(E2E_JOB_POLICY.prepareNoBuild);
 
-export const PREPARE_E2E_TRUSTED_BUILD_JOBS = new Set<string>(E2E_JOB_CONTRACT.prepareTrustedBuild);
+export const PREPARE_E2E_TRUSTED_BUILD_JOBS = new Set<string>(E2E_JOB_POLICY.prepareTrustedBuild);
 
 type WorkflowRecord = Record<string, unknown>;
 type WorkflowStep = WorkflowRecord & {
