+++ tools/e2e/workflow-boundary-policy.mts
@@ -37,7 +37,7 @@ export const E2E_ACTION_PROVENANCE = {
   },
 } as const;
 
-export const E2E_JOB_CONTRACT = {
+export const E2E_JOB_POLICY = {
   cliArtifactProducer: "generate-matrix",
   prepareNoBuild: [
     "bootstrap-install-smoke",
@@ -50,70 +50,4 @@ export const E2E_JOB_CONTRACT = {
     "whatsapp-qr-compact",
   ],
   prepareTrustedBuild: ["managed-image-protected-runtime"],
-  cliArtifactConsumers: [
-    "agent-turn-latency",
-    "bedrock-runtime-compatible-anthropic",
-    "brave-search",
-    "channels-add-remove",
-    "channels-stop-start",
-    "cloud-inference",
-    "cloud-onboard",
-    "common-egress-agent",
-    "concurrent-gateway-ports",
-    "cron-preflight-inference-local",
-    "dashboard-remote-bind",
-    "device-auth-health",
-    "double-onboard",
-    "full-e2e",
-    "gateway-guard-recovery",
-    "gpu-double-onboard",
-    "gpu-e2e",
-    "hermes-discord",
-    "hermes-e2e",
-    "hermes-gpu-startup",
-    "hermes-inference-switch",
-    "hermes-shields-config",
-    "hermes-slack",
-    "inference-routing",
-    "issue-2478-crash-loop-recovery",
-    "issue-4434-tui-unreachable-inference",
-    "issue-4462-scope-upgrade-approval",
-    "jetson-nvmap-gpu",
-    "kimi-inference-compat",
-    "live",
-    "llama-cpp-generic-gpu",
-    "mcp-bridge",
-    "mcp-bridge-dev",
-    "messaging-compatible-endpoint",
-    "messaging-providers",
-    "model-router-provider-routed-inference",
-    "network-policy",
-    "onboard-repair",
-    "onboard-resume",
-    "openclaw-discord-pairing",
-    "openclaw-inference-switch",
-    "openclaw-plugin-runtime-exdev",
-    "openclaw-plugin-runtime-exdev-release",
-    "openclaw-skill-cli",
-    "openclaw-slack-pairing",
-    "openclaw-tui-chat-correlation",
-    "openshell-credential-generation-window",
-    "openshell-gateway-auth-contract",
-    "openshell-gateway-upgrade",
-    "overlayfs-autofix",
-    "rebuild-hermes",
-    "rebuild-hermes-stale-base",
-    "rebuild-openclaw",
-    "retired-selector-compatibility",
-    "sandbox-operations",
-    "sandbox-survival",
-    "security-posture",
-    "sessions-agents-cli",
-    "shared-e2e",
-    "skill-agent",
-    "state-backup-restore",
-    "telegram-injection",
-    "token-rotation",
-    "tunnel-lifecycle",
-  ],
 } as const;
